# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Autoloker API
  version: 1.0.0
  description: API documentation for mobile authentication, profile editing, security work history CRUD, security certificate CRUD, public job vacancy and training listings, company- and BUJP-owned job vacancy and training management, job vacancy and training applications, paginated vacancy, published training, registered training listings, owned application status follow-up, active subscription plans, manual subscription order payments and proof uploads, bank references, and hierarchical location references.
  license:
    name: Proprietary
servers:
  - url: http://localhost
    description: Local development server
paths:
...
  /api/user-order-manuals:
    get:
      summary: List the authenticated user's manual subscription orders
      description: >-
        Returns the authenticated user's manual subscription order history, newest first, with five
        orders per page. Payment, verification, and expiration fields that are not yet set are
        returned as `null`, not empty strings. Each order includes the selected subscription plan.
      operationId: listUserOrderManuals
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: One-based pagination page number.
          schema:
            type: integer
            minimum: 1
            example: 1
          example: 1
      responses:
        '200':
          description: Manual subscription order history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOrderManualListResponse'
              examples:
                verifiedOrderHistory:
                  summary: Verified manual subscription order history
                  value:
                    status: true
                    data:
                      user_order_manuals:
                        - id: 1
                          uuid: 4c6f5d8a-0c2d-4f4b-a4b6-1f5a0e2c7d91
                          user_id: 12
                          master_subscription_id: 2
                          role: security
                          order_number: ORD-20260924-00001
                          price: '49000.00'
                          payment_method: transfer
                          payment_account: 'BCA 1234567890 a.n. PT Example'
                          payment_date: '2026-09-24T03:15:00.000000Z'
                          file: payment_ORD-20260924-00001_1727147700.jpg
                          status: approved
                          notes: Payment verified by the administrator.
                          rejected_reason: null
                          verified_by: '7'
                          verified_at: '2026-09-24T04:00:00.000000Z'
                          expired_at: '2026-10-24T03:15:00.000000Z'
                          created_at: '2026-09-24T03:00:00.000000Z'
                          updated_at: '2026-09-24T04:00:00.000000Z'
                          subscription:
                            id: 2
                            uuid: 8b1f4f8a-5d6f-4a4d-a43d-5f3c9f8c2a11
                            name: Profesional
                            slug: profesional
                            role: security
                            price: '49000.00'
                            duration: 1
                            duration_type: month
                            description: Untuk satpam aktif yang ingin melamar lebih banyak.
                            features:
                              job_vacancy:
                                enabled: true
                                limit: 5
                              training:
                                enabled: true
                                limit: 5
                            is_active: true
                            sort_order: 1
                      pagination:
                        current_page: 1
                        last_page: 1
                        per_page: 5
                        total: 1
                        from: 1
                        to: 1
                        has_more: false
                        next_page_url: null
                        previous_page_url: null
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      summary: Create a manual subscription order
      description: >-
        Creates a pending manual subscription order for the authenticated user's role. Payment and
        verification fields are `null` until the user submits proof and an administrator reviews
        the payment. A `pending_payment` order blocks creation of another order for this account,
        regardless of the selected plan. A conflict returns HTTP 409 with the existing order details.
      operationId: createUserOrderManual
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserOrderManualStoreRequest'
            example:
              subscription_uuid: 8b1f4f8a-5d6f-4a4d-a43d-5f3c9f8c2a11
      responses:
        '201':
          description: Manual subscription order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOrderManualResponse'
              examples:
                createdPendingOrder:
                  summary: Newly created order awaiting payment
                  value:
                    status: true
                    message: Pesanan paket langganan berhasil dibuat.
                    data:
                      user_order_manual:
                        id: 2
                        uuid: 9d2e4f6a-6c5b-4a2d-9f3e-7b1c8d0a5e64
                        user_id: 12
                        master_subscription_id: 2
                        role: security
                        order_number: ORD-20260926031500-AB12C
                        price: '49000.00'
                        payment_method: null
                        payment_account: null
                        payment_date: null
                        file: null
                        status: pending_payment
                        notes: null
                        rejected_reason: null
                        verified_by: null
                        verified_at: null
                        expired_at: null
                        created_at: '2026-09-26T03:15:00.000000Z'
                        updated_at: '2026-09-26T03:15:00.000000Z'
                        subscription:
                          id: 2
                          uuid: 8b1f4f8a-5d6f-4a4d-a43d-5f3c9f8c2a11
                          name: Profesional
                          slug: profesional
                          role: security
                          price: '49000.00'
                          duration: 1
                          duration_type: month
                          description: Untuk satpam aktif yang ingin melamar lebih banyak.
                          features:
                            job_vacancy:
                              enabled: true
                              limit: 5
                            training:
                              enabled: true
                              limit: 5
                          is_active: true
                          sort_order: 1
        '409':
          description: The authenticated user already has an order awaiting payment.
          content:
            application/json:
              schema:
                type: object
                required: [status, code, message, data]
                properties:
                  status:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: pending_payment_order_exists
                  message:
                    type: string
                    example: Masih ada pesanan paket yang menunggu pembayaran. Selesaikan atau batalkan pesanan tersebut sebelum membuat pesanan baru.
                  data:
                    type: object
                    required: [user_order_manual]
                    properties:
                      user_order_manual:
                        $ref: '#/components/schemas/UserOrderManual'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/user-order-manuals/{uuid}:
    get:
      summary: Get a manual subscription order
      description: >-
        Returns an order owned by the authenticated user with its subscription details and current
        payment status. Repeat this request to poll for status changes. Every response includes
        `Cache-Control: no-store` so a cached response does not mask an updated payment status.
      operationId: getUserOrderManual
      security:
        - bearerAuth: []
      parameters:
        - name: uuid
          in: path
          required: true
          description: UUID of the manual order.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Manual order details with its latest payment status
          headers:
            Cache-Control:
              description: Prevents caching the payment status response.
              schema:
                type: string
                example: no-store, private
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOrderManualResponse'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          description: Order does not exist or is not owned by the authenticated user
          headers:
            Cache-Control:
              description: Prevents caching the payment status response.
              schema:
                type: string
                example: no-store, private
          content:
            application/json:
              schema:
                type: object
                required: [status, message, data]
                properties:
                  status:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: Pesanan tidak ditemukan.
                  data:
                    type: 'null'
...
components:
  schemas:
...
    TrainingPagination:
      type: object
      required:
        - current_page
        - last_page
        - per_page
        - total
        - from
        - to

        - has_more
        - next_page_url
        - previous_page_url
      properties:
        current_page:
          type: integer
          example: 1
        last_page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 5
        total:
          type: integer
          example: 1
        from:
          type: [integer, 'null']
          example: 1
        to:
          type: [integer, 'null']
          example: 1
        has_more:
          type: boolean
          example: false
        next_page_url:
          type: [string, 'null']
          format: uri
          example: null
        previous_page_url:
          type: [string, 'null']
          format: uri
          example: null
...
    UserOrderManualStoreRequest:
      type: object
      additionalProperties: false
      required: [subscription_uuid]
      properties:
        subscription_uuid:
          type: string
          format: uuid
          description: UUID of an active subscription plan for the authenticated user's role.
...
    UserOrderManualListResponse:
      type: object
      required: [status, data]
      properties:
        status:
          type: boolean
          example: true
        data:
          type: object
          required: [user_order_manuals, pagination]
          properties:
            user_order_manuals:
              type: array
              items:
                $ref: '#/components/schemas/UserOrderManual'
            pagination:
              $ref: '#/components/schemas/TrainingPagination'
    UserOrderManualResponse:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          required: [user_order_manual]
          properties:
            user_order_manual:
              $ref: '#/components/schemas/UserOrderManual'
    UserOrderManualSubscription:
      type: object
      description: Subscription plan snapshot embedded in a manual order.
      required: [id, uuid, name, slug, role, price, duration, duration_type, description, features, is_active, sort_order]
      properties:
        id:
          type: integer
          example: 2
        uuid:
          type: [string, 'null']
          format: uuid
          example: 8b1f4f8a-5d6f-4a4d-a43d-5f3c9f8c2a11
        name:
          type: [string, 'null']
          example: Profesional
        slug:
          type: [string, 'null']
          example: profesional
        role:
          type: [string, 'null']
          enum: [security, bujp, client, null]
        price:
          type: string
          description: Subscription price snapshot as a decimal string with two fractional digits.
          example: '49000.00'
        duration:
          type: integer
          example: 1
        duration_type:
          type: string
          enum: [day, month, year]
          example: month
        description:
          type: [string, 'null']
          example: Untuk satpam aktif yang ingin melamar lebih banyak.
        features:
          type: [object, 'null']
          additionalProperties: true
          example:
            job_vacancy:
              enabled: true
              limit: 5
            training:
              enabled: true
              limit: 5
        is_active:
          type: boolean
          example: true
        sort_order:
          type: integer
          minimum: 0
          example: 1
    UserOrderManual:
      type: object
      required:
        - id
        - uuid
        - user_id
        - master_subscription_id
        - role
        - order_number
        - price
        - payment_method
        - payment_account
        - payment_date
        - file
        - status
        - notes
        - rejected_reason
        - verified_by
        - verified_at
        - expired_at
        - created_at
        - updated_at
        - subscription
      properties:
        id:
          type: integer
          example: 1
        uuid:
          type: string
          format: uuid
          example: 4c6f5d8a-0c2d-4f4b-a4b6-1f5a0e2c7d91
        user_id:
          type: integer
          example: 12
        master_subscription_id:
          type: integer
          example: 2
        role:
          type: string
          enum: [security, bujp, client]
        order_number:
          type: string
          example: ORD-20260924-00001
        price:
          type: string
          description: Order price as a decimal string with two fractional digits.
          example: '49000.00'
        payment_method:
          type: [string, 'null']
          example: transfer
        payment_account:
          type: [string, 'null']
          example: 'BCA 1234567890 a.n. PT Example'
        payment_date:
          type: [string, 'null']
          format: date-time
          example: '2026-09-24T03:15:00.000000Z'
        file:
          type: [string, 'null']
          description: Stored payment proof filename, when uploaded.
          example: payment_ORD-20260924-00001_1727147700.jpg
        status:
          type: string
          description: 'Order lifecycle status: pending_payment, verification, approved, rejected, or cancelled.'
        notes:
          type: [string, 'null']
          example: Pembayaran sedang ditinjau oleh administrator.
        rejected_reason:
          type: [string, 'null']
          example: null
        verified_by:
          type: [string, 'null']
          description: Authenticated administrator user ID that reviewed the order.
          example: null
        verified_at:
          type: [string, 'null']
          format: date-time
          example: null
        expired_at:
          type: [string, 'null']
          format: date-time
          example: null
        created_at:
          type: [string, 'null']
          format: date-time
          example: '2026-09-24T03:00:00.000000Z'
        updated_at:
          type: [string, 'null']
          format: date-time
          example: '2026-09-24T03:00:00.000000Z'
        subscription:
          anyOf:
            - $ref: '#/components/schemas/UserOrderManualSubscription'
            - type: 'null'
...
    ValidationErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    ErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          example: false
        message:
          type: string
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    Unauthenticated:
      description: Missing, invalid, expired, or revoked access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: The requested resource or security profile was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque token
...